chore(alloy): cleanup hardforks
