--- crates/alloy/hardforks/src/base/mainnet.rs
@@ -1,22 +1,0 @@
-//! Base Mainnet hardfork starting points
-
-use crate::optimism::mainnet::*;
-
-/// Bedrock base hardfork activation block is 0.
-pub const BASE_MAINNET_BEDROCK_BLOCK: u64 = 0;
-/// Regolith base hardfork activation timestamp is 0.
-pub const BASE_MAINNET_REGOLITH_TIMESTAMP: u64 = OP_MAINNET_REGOLITH_TIMESTAMP;
-/// Canyon base hardfork activation timestamp is 1704992401.
-pub const BASE_MAINNET_CANYON_TIMESTAMP: u64 = OP_MAINNET_CANYON_TIMESTAMP;
-/// Ecotone base hardfork activation timestamp is 1710374401.
-pub const BASE_MAINNET_ECOTONE_TIMESTAMP: u64 = OP_MAINNET_ECOTONE_TIMESTAMP;
-/// Fjord base hardfork activation timestamp is 1720627201.
-pub const BASE_MAINNET_FJORD_TIMESTAMP: u64 = OP_MAINNET_FJORD_TIMESTAMP;
-/// Granite base hardfork activation timestamp is 1726070401.
-pub const BASE_MAINNET_GRANITE_TIMESTAMP: u64 = OP_MAINNET_GRANITE_TIMESTAMP;
-/// Holocene base hardfork activation timestamp is 1736445601.
-pub const BASE_MAINNET_HOLOCENE_TIMESTAMP: u64 = OP_MAINNET_HOLOCENE_TIMESTAMP;
-/// Isthmus base hardfork activation timestamp is 1746806401.
-pub const BASE_MAINNET_ISTHMUS_TIMESTAMP: u64 = OP_MAINNET_ISTHMUS_TIMESTAMP;
-/// Jovian base hardfork activation timestamp is `1_763_481_601`.
-pub const BASE_MAINNET_JOVIAN_TIMESTAMP: u64 = OP_MAINNET_JOVIAN_TIMESTAMP;
