--- crates/alloy/hardforks/src/base/mod.rs
@@ -1,6 +1,0 @@
-//! Base hardfork starting points
-
-pub mod mainnet;
-pub use mainnet::*;
-pub mod sepolia;
-pub use sepolia::*;
