--- crates/alloy/hardforks/src/base/sepolia.rs
@@ -1,22 +1,0 @@
-//! Base Sepolia hardfork starting points
-
-use crate::optimism::sepolia::*;
-
-/// Bedrock base sepolia hardfork activation block is 0.
-pub const BASE_SEPOLIA_BEDROCK_BLOCK: u64 = OP_SEPOLIA_BEDROCK_BLOCK;
-/// Regolith base sepolia hardfork activation timestamp is 0.
-pub const BASE_SEPOLIA_REGOLITH_TIMESTAMP: u64 = OP_SEPOLIA_REGOLITH_TIMESTAMP;
-/// Canyon base sepolia hardfork activation timestamp is 1699981200.
-pub const BASE_SEPOLIA_CANYON_TIMESTAMP: u64 = OP_SEPOLIA_CANYON_TIMESTAMP;
-/// Ecotone base sepolia hardfork activation timestamp is 1708534800.
-pub const BASE_SEPOLIA_ECOTONE_TIMESTAMP: u64 = OP_SEPOLIA_ECOTONE_TIMESTAMP;
-/// Fjord base sepolia hardfork activation timestamp is 1716998400.
-pub const BASE_SEPOLIA_FJORD_TIMESTAMP: u64 = OP_SEPOLIA_FJORD_TIMESTAMP;
-/// Granite base sepolia hardfork activation timestamp is 1723478400.
-pub const BASE_SEPOLIA_GRANITE_TIMESTAMP: u64 = OP_SEPOLIA_GRANITE_TIMESTAMP;
-/// Holocene base sepolia hardfork activation timestamp is 1732633200.
-pub const BASE_SEPOLIA_HOLOCENE_TIMESTAMP: u64 = OP_SEPOLIA_HOLOCENE_TIMESTAMP;
-/// Isthmus base sepolia hardfork activation timestamp is 1744905600.
-pub const BASE_SEPOLIA_ISTHMUS_TIMESTAMP: u64 = OP_SEPOLIA_ISTHMUS_TIMESTAMP;
-/// Jovian base sepolia hardfork activation timestamp is `1_762_963_201`.
-pub const BASE_SEPOLIA_JOVIAN_TIMESTAMP: u64 = OP_SEPOLIA_JOVIAN_TIMESTAMP;
