+++ crates/alloy/hardforks/src/mainnet.rs
@@ -1,0 +1,18 @@
+/// Bedrock base hardfork activation block is 0.
+pub const BASE_MAINNET_BEDROCK_BLOCK: u64 = 0;
+/// Regolith base hardfork activation timestamp is 0.
+pub const BASE_MAINNET_REGOLITH_TIMESTAMP: u64 = 0;
+/// Canyon base hardfork activation timestamp is 1704992401.
+pub const BASE_MAINNET_CANYON_TIMESTAMP: u64 = 1_704_992_401;
+/// Ecotone base hardfork activation timestamp is 1710374401.
+pub const BASE_MAINNET_ECOTONE_TIMESTAMP: u64 = 1_710_374_401;
+/// Fjord base hardfork activation timestamp is 1720627201.
+pub const BASE_MAINNET_FJORD_TIMESTAMP: u64 = 1_720_627_201;
+/// Granite base hardfork activation timestamp is 1726070401.
+pub const BASE_MAINNET_GRANITE_TIMESTAMP: u64 = 1_726_070_401;
+/// Holocene base hardfork activation timestamp is 1736445601.
+pub const BASE_MAINNET_HOLOCENE_TIMESTAMP: u64 = 1_736_445_601;
+/// Isthmus base hardfork activation timestamp is 1746806401.
+pub const BASE_MAINNET_ISTHMUS_TIMESTAMP: u64 = 1_746_806_401;
+/// Jovian base hardfork activation timestamp is 1764691201.
+pub const BASE_MAINNET_JOVIAN_TIMESTAMP: u64 = 1_764_691_201;
