--- crates/alloy/hardforks/src/optimism/mainnet.rs
@@ -1,24 +1,0 @@
-//! Optimism Mainnet hardfork starting points
-
-//------------------------ OVM chain ------------------------//
-/// Berlin hardfork activation block is 3950000.
-pub const OP_MAINNET_BERLIN_BLOCK: u64 = 3_950_000;
-//------------------------ EVM chain ------------------------//
-/// Bedrock hardfork activation block is 105235063.
-pub const OP_MAINNET_BEDROCK_BLOCK: u64 = 105_235_063;
-/// Regolith hardfork activation timestamp is 0.
-pub const OP_MAINNET_REGOLITH_TIMESTAMP: u64 = 0;
-/// Canyon hardfork activation timestamp is 1704992401.
-pub const OP_MAINNET_CANYON_TIMESTAMP: u64 = 1_704_992_401;
-/// Ecotone hardfork activation timestamp is 1710374401.
-pub const OP_MAINNET_ECOTONE_TIMESTAMP: u64 = 1_710_374_401;
-/// Fjord hardfork activation timestamp is 1720627201.
-pub const OP_MAINNET_FJORD_TIMESTAMP: u64 = 1_720_627_201;
-/// Granite hardfork activation timestamp is 1726070401.
-pub const OP_MAINNET_GRANITE_TIMESTAMP: u64 = 1_726_070_401;
-/// Holocene hardfork activation timestamp is 1736445601.
-pub const OP_MAINNET_HOLOCENE_TIMESTAMP: u64 = 1_736_445_601;
-/// Isthmus hardfork activation timestamp is 1746806401.
-pub const OP_MAINNET_ISTHMUS_TIMESTAMP: u64 = 1_746_806_401;
-/// Jovian hardfork activation timestamp is `1_764_691_201` # Tue 2 Dec 2025 16:00:01 UTC
-pub const OP_MAINNET_JOVIAN_TIMESTAMP: u64 = 1_764_691_201;
