--- crates/alloy/hardforks/src/optimism/mod.rs
@@ -1,6 +1,0 @@
-//! Optimism hardfork starting points
-
-pub mod mainnet;
-pub use mainnet::*;
-pub mod sepolia;
-pub use sepolia::*;
