--- crates/alloy/hardforks/src/optimism/sepolia.rs
@@ -1,20 +1,0 @@
-//! Optimism Sepolia hardfork starting points
-
-/// Bedrock sepolia hardfork activation block is 0.
-pub const OP_SEPOLIA_BEDROCK_BLOCK: u64 = 0;
-/// Regolith sepolia hardfork activation timestamp is 0.
-pub const OP_SEPOLIA_REGOLITH_TIMESTAMP: u64 = 0;
-/// Canyon sepolia hardfork activation timestamp is 1699981200.
-pub const OP_SEPOLIA_CANYON_TIMESTAMP: u64 = 1_699_981_200;
-/// Ecotone sepolia hardfork activation timestamp is 1708534800.
-pub const OP_SEPOLIA_ECOTONE_TIMESTAMP: u64 = 1_708_534_800;
-/// Fjord sepolia hardfork activation timestamp is 1716998400.
-pub const OP_SEPOLIA_FJORD_TIMESTAMP: u64 = 1_716_998_400;
-/// Granite sepolia hardfork activation timestamp is 1723478400.
-pub const OP_SEPOLIA_GRANITE_TIMESTAMP: u64 = 1_723_478_400;
-/// Holocene sepolia hardfork activation timestamp is 1732633200.
-pub const OP_SEPOLIA_HOLOCENE_TIMESTAMP: u64 = 1_732_633_200;
-/// Isthmus sepolia hardfork activation timestamp is 1744905600.
-pub const OP_SEPOLIA_ISTHMUS_TIMESTAMP: u64 = 1_744_905_600;
-/// Jovian sepolia hardfork activation timestamp is `1_763_568_001` # Wed 19 Nov 2025 16:00:01 UTC.
-pub const OP_SEPOLIA_JOVIAN_TIMESTAMP: u64 = 1_763_568_001;
