+++ crates/alloy/hardforks/src/sepolia.rs
@@ -1,0 +1,18 @@
+/// Bedrock base sepolia hardfork activation block is 0.
+pub const BASE_SEPOLIA_BEDROCK_BLOCK: u64 = 0;
+/// Regolith base sepolia hardfork activation timestamp is 0.
+pub const BASE_SEPOLIA_REGOLITH_TIMESTAMP: u64 = 0;
+/// Canyon base sepolia hardfork activation timestamp is 1699981200.
+pub const BASE_SEPOLIA_CANYON_TIMESTAMP: u64 = 1_699_981_200;
+/// Ecotone base sepolia hardfork activation timestamp is 1708534800.
+pub const BASE_SEPOLIA_ECOTONE_TIMESTAMP: u64 = 1_708_534_800;
+/// Fjord base sepolia hardfork activation timestamp is 1716998400.
+pub const BASE_SEPOLIA_FJORD_TIMESTAMP: u64 = 1_716_998_400;
+/// Granite base sepolia hardfork activation timestamp is 1723478400.
+pub const BASE_SEPOLIA_GRANITE_TIMESTAMP: u64 = 1_723_478_400;
+/// Holocene base sepolia hardfork activation timestamp is 1732633200.
+pub const BASE_SEPOLIA_HOLOCENE_TIMESTAMP: u64 = 1_732_633_200;
+/// Isthmus base sepolia hardfork activation timestamp is 1744905600.
+pub const BASE_SEPOLIA_ISTHMUS_TIMESTAMP: u64 = 1_744_905_600;
+/// Jovian base sepolia hardfork activation timestamp is 1763568001.
+pub const BASE_SEPOLIA_JOVIAN_TIMESTAMP: u64 = 1_763_568_001;
